rename events reducer file, make event reducer file
