+++ src/redux/reducers/event.reducer.js
@@ -1,20 +1,20 @@
 import { ACTION_TYPES } from '../../constants/ActionTypes';
 
 const DEFAULT_STATE = {
-  events: null,
+  event: null,
   isFetching: false,
   didInvalidate: false,
 };
 
-export const eventsReducer = (state = DEFAULT_STATE, action) => {
+export const eventReducer = (state = DEFAULT_STATE, action) => {
 
   switch(action.type) {
-    case ACTION_TYPES.GET_EVENTS:
+    case ACTION_TYPES.GET_EVENT:
       return {...state, isFetching: true};
-    case ACTION_TYPES.SET_EVENTS:
-      return {...state, events: action.payload.events, isFetching: false, didInvalidate: false}
+    case ACTION_TYPES.SET_EVENT:
+      return {...state, event: action.payload.event, isFetching: false, didInvalidate: false}
     case ACTION_TYPES.SET_EVENT_INVALIDATE:
-      return {...state, isFetching: false, didInvalidate: true, events: null}
+      return {...state, isFetching: false, didInvalidate: true, event: null}
     default:
       return state;
   }
